ScriptCache: Respect run_script's calling convention
Although if we make the whoel sha'ing thing totally opaque to the user,
i.e., they never know this is going on, then Lua script == "command"

If that is the case, the the caller only needs to know all the
"commands" that are available. That might be neat.

Might push that as follow-up commits as a study of what that might look
like.